Do not always allow closing via keyboard.
Apparently, workspaces can be active or not active. I might be looking
at a workspace, but it might not be active for uhhhh reasons?

Fixing this mess is best left for a gnome-shell patch, so for now I'll
just emulate the behaviour of the Enter key when selecting a window. It
makes no sense in my opinion, but I'd rather not risk accidentally
closing random windows in other workspaces (Ask me how I know).